Unify CSP/HSM disabling tests
